abstracted the functions in backslash.test.ts
